+++ src/stream/MediaStream.h
@@ -10,7 +10,7 @@ public:
     bool isValid();
     size_t write(uint8_t byte) override;
 
-    virtual size_t totalSize() = 0;
+    virtual int totalSize() = 0;
 
 protected:
     void setValid(bool valid);
